Calculate weightings on a per disease basis, and group model prep tasks into transactional ModelRunGateKeeper
