+++ src/DataAcquisition/src/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/config/beans.xml
@@ -58,9 +58,8 @@
     <!-- Post-quality control beans -->
     <bean id="postQcManager" class="uk.ac.ox.zoo.seeg.abraid.mp.dataacquisition.qc.PostQCManager" autowire="constructor"/>
 
-    <!-- Disease extent generation beans -->
+    <!-- Disease extent generation bean -->
     <bean id="diseaseExtentGenerator" class="uk.ac.ox.zoo.seeg.abraid.mp.dataacquisition.diseaseextent.DiseaseExtentGenerator" autowire="constructor"/>
-    <bean id="singleDiseaseExtentGenerator" class="uk.ac.ox.zoo.seeg.abraid.mp.dataacquisition.diseaseextent.SingleDiseaseExtentGenerator" autowire="constructor"/>
 
     <bean id="configurationService" class="uk.ac.ox.zoo.seeg.abraid.mp.dataacquisition.config.ConfigurationServiceImpl">
         <constructor-arg name="basicProperties" value="${internal.properties.fileLocation}"/>
@@ -71,6 +70,9 @@
 
     <!-- Model run requesting beans -->
     <bean id="modelRunRequester" class="uk.ac.ox.zoo.seeg.abraid.mp.dataacquisition.model.ModelRunRequester" autowire="constructor"/>
+
+    <!-- Model run preparation tasks beans -->
+    <bean id="modelRunGateKeeper" class="uk.ac.ox.zoo.seeg.abraid.mp.dataacquisition.model.ModelRunGateKeeper" autowire="constructor"/>
 
     <!-- Main bean (program entry point) -->
     <bean id="main" class="uk.ac.ox.zoo.seeg.abraid.mp.dataacquisition.Main" autowire="constructor"/>
