+++ src/DataAcquisition/src/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/config/dataacquisition.internal.properties
@@ -2,4 +2,4 @@
 # Copyright (c) 2014 University of Oxford
 
 # The date on which DiseaseOccurrenceReviews were last retrieved.
-lastRetrievalDate=
+lastRetrievalDate=2014-05-16T17:22:53.721
