+++ src/DataAcquisition/src/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/weightings/WeightingsCalculator.java
@@ -25,9 +25,10 @@ import static ch.lambdaj.Lambda.*;
 public class WeightingsCalculator {
     private static final Logger LOGGER = Logger.getLogger(WeightingsCalculator.class);
     private static final String STARTING_UPDATE_WEIGHTINGS =
-            "Starting process to update occurrences' expert weightings";
+            "Starting process to update occurrences' expert weightings for disease %d";
     private static final String NOT_UPDATING_WEIGHTINGS =
-            "Occurrences' expert weightings will not be updated - a week has not elapsed since last retrieval on %s";
+            "Occurrences' expert weightings will not be updated for disease %d" +
+                    " - a week has not elapsed since last retrieval on %s";
     private static final String NO_NEW_REVIEWS = "No new reviews have been submitted - weightings will not be updated";
     private static final String UPDATING_WEIGHTINGS =
             "Updating expert weightings for %d disease occurrences given %d new reviews";
@@ -44,21 +45,22 @@ public class WeightingsCalculator {
      * For every disease occurrence point that has had new reviews submitted since the last recalculation a week ago,
      * calculate its new weighting, by taking the weighted average of every expert review in the database (not just
      * the new reviews).
+     * @param diseaseGroupId The id of the disease group.
      */
     @Transactional
-    public void updateDiseaseOccurrenceWeightings() {
+    public void updateDiseaseOccurrenceWeightings(int diseaseGroupId) {
         LocalDateTime lastRetrievalDate = configurationService.getLastRetrievalDate();
         if (shouldContinue(lastRetrievalDate)) {
-            LOGGER.info(STARTING_UPDATE_WEIGHTINGS);
+            LOGGER.info(String.format(STARTING_UPDATE_WEIGHTINGS, diseaseGroupId));
             configurationService.setLastRetrievalDate(LocalDateTime.now());
-            List<DiseaseOccurrenceReview> allReviews = getAllReviews(lastRetrievalDate);
+            List<DiseaseOccurrenceReview> allReviews = getAllReviewsForDiseaseGroup(lastRetrievalDate, diseaseGroupId);
             if (allReviews.isEmpty()) {
                 LOGGER.info(NO_NEW_REVIEWS);
             } else {
                 calculateNewDiseaseOccurrenceWeightings(allReviews);
             }
         } else {
-            LOGGER.info(String.format(NOT_UPDATING_WEIGHTINGS, lastRetrievalDate.toString()));
+            LOGGER.info(String.format(NOT_UPDATING_WEIGHTINGS, diseaseGroupId, lastRetrievalDate.toString()));
         }
     }
 
@@ -73,11 +75,13 @@ public class WeightingsCalculator {
         }
     }
 
-    private List<DiseaseOccurrenceReview> getAllReviews(LocalDateTime lastRetrievalDate) {
+    private List<DiseaseOccurrenceReview> getAllReviewsForDiseaseGroup(LocalDateTime lastRetrievalDate,
+                                                                       int diseaseGroupId) {
         if (lastRetrievalDate == null) {
-            return diseaseService.getAllDiseaseOccurrenceReviews();
+            return diseaseService.getAllDiseaseOccurrenceReviewsByDiseaseGroupId(diseaseGroupId);
         } else {
-            return diseaseService.getAllReviewsForDiseaseOccurrencesWithNewReviewsSinceLastRetrieval(lastRetrievalDate);
+            return diseaseService.getAllReviewsForDiseaseGroupOccurrencesWithNewReviewsSinceLastRetrieval(
+                    lastRetrievalDate, diseaseGroupId);
         }
     }
